Add BottomNavigationViewController.
